requirements.txt
jsonpath_ng	    ==1.6.1
python-dotenv	    ==1.0.1
SpeechRecognition   == 3.10.4
opencv-python       == 4.10.0.84
ultralytics         == 8.2.76
omegaconf           == 2.3.0
fastapi             == 0.111.1
PyAudio             == 0.2.14
pyttsx3             == 2.91
PyQt6               == 6.7.0
PyQt6-WebEngine     == 6.7.0
llama_index==0.11.5
llama_parse==0.5.2

